Remove Preset Pages (#19)
# Description
The `pages` directory contained content that was made specifically for @Behemyth and doesn't need to be in the shared codebase. The pages and routes for each user of this website are going to be different. Therefor we now use dynamic, nested routes which are prepended by `_`. See the [documentation](https://nuxtjs.org/guides/features/file-system-routing#dynamic-nested-routes) for more details

Closes #18 

# Changes
- The deprecated `v-content` component from Vuetify has been replaced with `v-main` which kinda does the same thing I guess.
- Small linting fixes. @zyteka 👼
- Removal of all the hard-coded pages found in the `pages` root directory. All the stub content prepended with `_` contains incomplete and potentially missing data.